requirements.txt
asgiref==3.4.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.21.7
idna==3.3
inflection==0.5.1
packaging==23.2
Pillow==8.3.2
psycopg2==2.9.1
PyJWT==2.3.0
python-dotenv==0.20.0
pytz==2021.3
PyYAML==6.0.1
requests==2.27.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
suds==1.0.0
uritemplate==4.1.1
urllib3==1.26.8
